refactor: use static imports for services
