+++ pages/CultivosPage.tsx
@@ -10,6 +10,7 @@ import Toast from '../components/Toast';
 import LeafIcon from '../components/icons/LeafIcon';
 import Loader from '../components/Loader';
 import ErrorBanner from '../components/ErrorBanner';
+import { getCultivos } from '../services/cultivoService';
 
 const CultivosPage: React.FC = () => {
   const { t } = useTranslation();
@@ -23,11 +24,9 @@ const CultivosPage: React.FC = () => {
     let mounted = true;
     async function fetchCultivos() {
       try {
-        const { getCultivos } = await import('../services/cultivoService');
         const data = await getCultivos();
         if (mounted) setCultivos(data);
       } catch {
-        // Error handled, variable not used
         setError(t('cultivosPage.error_loading'));
       } finally {
         if (mounted) setLoading(false);
